requirements.txt
Whoosh
aldryn-search
django-cms-articles
djangocms-bootstrap5
leprikon
